Prepare streamlit for Heroku

+++ streamlit/utils/sound_utils.py
@@ -1,3 +1,9 @@
+"""
+sound utils
+-----------
+
+Functions useful for processing sounds
+"""
 import sys
 from pathlib import Path, PurePath
 
@@ -15,19 +21,22 @@ class SoundHelper:
     # Filename template
     file_name = ""
 
-    def __init__(self):
-        # Max number of files to keep
-        self.rotate_files = 50
-        self.current_file = 0
-        try:
-            Path(self.sounds_path).mkdir(parents=True, exist_ok=True)
-        except FileExistsError:
-            pass
-        finally:
-            # Empty folder
-            [f.unlink() for f in Path(self.sounds_path).glob("*") if f.is_file()]
+    def __init__(self, mode="wav"):
+        """Prepares the folder to store the files"""
+        if mode == "wav":
+            # Max number of files to keep
+            self.rotate_files = 50
+            self.current_file = 0
+            try:
+                Path(self.sounds_path).mkdir(parents=True, exist_ok=True)
+            except FileExistsError:
+                pass
+            finally:
+                # Empty folder
+                [f.unlink() for f in Path(self.sounds_path).glob("*") if f.is_file()]
 
     def save_sound(self, buffer):
+        """Stores and queue a pydub AudioSegment object as sound file"""
         self.file_name = (
             f"{self.sounds_path}/sound{self.current_file}.{self.file_format}"
         )
@@ -46,6 +55,7 @@ class SoundHelper:
         return self.file_name
 
     def stream_sound(self, buffer):
+        """Queue pydub AudioSegment file"""
         print(
             f"Buffer received ({buffer.duration_seconds}):\nChannels: {buffer.channels}\nSample Width: {buffer.sample_width}, \nFrame rate: {buffer.frame_rate}"
         )
